UPD 9.2
* features for new post
* ability to registration on events

//focus on myevents switchers, ok?

+++ pages/news.php
@@ -47,9 +47,52 @@ style="position:relative; margin:auto auto; width:30; height:30;"
 
 </form>
 
+<form id="registr" method="post" style="dispay:none">
+	<input id="toRegistr" style="display:none" name="toRegistr"></input>
+	<input id="toDeregistr" style="display:none" name="toDeregistr"></input>
+</form>
+
+<!-- registrating user on event -->
+<?php  include "/php/registration.php"?>
+<?php
+	if($_POST["toRegistr"]!=null)
+	{
+		$newsID=$_POST["toRegistr"];
+		$userID=getUserValue($_COOKIE["userID"],"ID");
+
+		$sqlCon= new mysqli("127.0.0.1:3306","root","","ITB");
+		$result=$sqlCon->query("INSERT INTO Visitors(UserID, NewsID) VALUES('$userID','$newsID')");
+		if($result==true)
+		{
+			echo "<script>gotoNews();</script>";
+		}
+		else
+		{
+			echo "<script>console.log(\"registrate failure\");;</script>";
+		}
+	}
+
+	if($_POST["toDeregistr"]!=null)
+	{
+		$newsID=$_POST["toDeregistr"];
+		$userID=getUserValue($_COOKIE["userID"],"ID");
+
+		$sqlCon= new mysqli("127.0.0.1:3306","root","","ITB");
+		$result=$sqlCon->query("DELETE FROM Visitors WHERE `UserID`='$userID' AND `NewsID`='$newsID'");
+		if($result==true)
+		{
+			echo "<script>gotoNews();</script>";
+		}
+		else
+		{
+			echo "<script>console.log(\"registrate failure\");;</script>";
+		}
+	}
+?>
+
 <!--билдинг новостей-->
 <?php  include "/php/newsShower.php"?>
-<?php  include "/php/registration.php"?>
+
 
 <form id="newsForm" type="get" action="showNews.php">
 <input name="newsID" id="newsID" type="text" style="display:none"></input>
@@ -63,9 +106,14 @@ style="position:relative; margin:auto auto; width:30; height:30;"
 	}
 </script>
 
-<form method="post" id="deleter">
-	<input type="text" style="display:none" id="toDelete" name="toDelete"></input>
-</form>
+<?php
+if(getUserStatus()=="Gold")
+{
+echo "<form method=\"post\" id=\"deleter\">";
+	echo "<input type=\"text\" style=\"display:none\" id=\"toDelete\" name=\"toDelete\"></input>";
+echo "</form>";
+}
+?>
 
 <script>
 	function deletePost(Id)
@@ -78,7 +126,7 @@ style="position:relative; margin:auto auto; width:30; height:30;"
 <?php
 $deleter=$_POST["toDelete"];
 
-if($deleter!=null)
+if($deleter!=null && getUserStatus()=="Gold")
 {
 	deleteNews($deleter);
 }
@@ -90,15 +138,20 @@ for($i=0;$i<count($statti);$i++)
 $zag=$statti[$i]->head;
 $cont=$statti[$i]->smallContent;
 $id=$statti[$i]->ID;
+$type=$statti[$i]->type;
+
+$bg=$statti[$i]->type=="Event"?"#3232aa":"#245eac";
+if($i==0 && $type=="Event") { $bg="#8209a3"; }
 
 echo "<div style=\" min-width:600; margin:20 auto; top:200;\" class=\"main\" id=\"main_$i\">";
 
-	//голвоа
-	echo "<div  style=\"cursor:pointer; background:#245eac; width:100%; height:35; box-shadow: 0 0 10px; \"> <p onclick=\"goNews($id);\" class=\"zagolovok\">$zag</p>";
-	//$k=getUserValue($_COOKIE["userID"],"Status");
+	//голова
+	echo "<div  style=\"cursor:pointer; background:$bg; width:100%; height:35; box-shadow: 0 0 10px; \"> <p onclick=\"goNews($id);\" class=\"zagolovok\">";
+	if ($i==0 && $type=="Event") { echo "$zag ( НАЙБЛИЖЧЕ )"; } else { echo "$zag"; }
+	echo "</p>";
 	$de=$_COOKIE["userID"];
 
-	if(getUserValue($_COOKIE["userID"],"Status")=="Gold")
+	if(getUserStatus()=="Gold")
 	{
 		echo "<div onclick=\"deletePost($id);\" id=\"deleteBtn\">";
 		echo "<div style=\"background:white; transform: rotate(45deg); margin:13 0%; position:absolute; width:100%; height:3;\"></div>";
@@ -111,15 +164,22 @@ echo "<div style=\" min-width:600; margin:20 auto; top:200;\" class=\"main\" id=
 	echo "<p style=\"margin:10 5;\" class=\"text_S\">$cont</p>";
 
 	//кнопки
+	
 	echo "<div style=\"position:absolute; bottom:5;\">";
 
 	echo "<div class=\"btnS\" onclick=\"goNews($id);\" style=\"width:200; margin:0 0; position:relative; display:inline-block; bottom:10; box-shadow: 0 0 10px; left:10; height:30; background:#245eac;\">";
 		echo "<a class=\"small_btn_text\" style=\"position:absolute; margin:0 40;\">більше . . .</a>";
 	echo "</div>";
-	if(getSqlValueById($id,"Type","News")=="Event")
+	if(getSqlValueById($id,"Type","News")=="Event" && isOnEvent(getUserValue($_COOKIE["userID"],"ID"),$id)==0)
 	{
 	echo "<div class=\"btnS\" onclick=\"\" style=\"width:275; margin:0 10; position:relative; display:inline-block; bottom:10; box-shadow: 0 0 10px; left:10; height:30; background:#245eac;\">";
-	echo "<a class=\"small_btn_text\" style=\"position:absolute; margin:0 40;\">зареєструватись</a>";
+	echo "<a class=\"small_btn_text\" onclick=\"registrateOnEvent($id);\" style=\"position:absolute; margin:0 40;\">зареєструватись</a>";
+	echo "</div>";
+	}
+	if(getSqlValueById($id,"Type","News")=="Event" && isOnEvent(getUserValue($_COOKIE["userID"],"ID"),$id)==1)
+	{
+	echo "<div class=\"btnS\" onclick=\"\" style=\"width:325; margin:0 10; position:relative; display:inline-block; bottom:10; box-shadow: 0 0 10px; left:10; height:30; background:#245eac;\">";
+	echo "<a class=\"small_btn_text\" onclick=\"deRegistrateOnEvent($id);\" style=\"position:absolute; margin:0 40;\">відмінити реєстрацію</a>";
 	echo "</div>";
 	}
 	echo "</div>";
@@ -127,10 +187,24 @@ echo "<div style=\" min-width:600; margin:20 auto; top:200;\" class=\"main\" id=
 	$AuthID=getSqlValueById($id, "Creator_ID","News");
 	$Auth=getSqlValueById($AuthID,"Name","Users");
 	$DateOfCreate=getSqlValueById($id,"CreateDate","News");
+	$StartDate=getSqlValueById($id,"StartDate","News");
+	$Type=getSqlValueById($id,"Type","News");
+	$Price=getSqlValueById($id, "Price","News");
+
+	$wrDate=strtotime($StartDate);
+
+	if($Type=="Event")
+	{
+		echo "<a class=\"text_S\" style=\"position:absolute; color:green; right:0; bottom:55; opacity:0.5;\">$Price баллів</a>";
+
+		$hour=date("H",$wrDate); $mins=date("i",$wrDate); $month=date("m",$wrDate); $day=date("d",$wrDate);
+		echo "<a class=\"text_S\" style=\"position:absolute; color:red; right:0; text-align:right; background:black; bottom:35; opacity:0.7;\">початок: $month/$day о $hour:$mins</a>";
+	}
 	echo "<a class=\"text_S\" style=\"position:absolute; right:0; bottom:15; opacity:0.5;\">$DateOfCreate</a>";
-	echo "<a class=\"text_S\" style=\"position:absolute; right:0; bottom:0; opacity:0.5;\">$Auth</a>";
+	echo "<a class=\"text_S\" style=\"position:absolute; right:0; text-align:right; bottom:0; opacity:0.5;\">$Auth</a>";
 
 echo "</div>";
+
 
 }
 ?>
